[python] refactor file names, fix import scope
