+++ .gitignore
@@ -1,3 +1,5 @@
 .Rproj.user
 .Rhistory
 .RData
+
+python/misc/*
