+++ python/dalex/dalex/__init__.py
@@ -1,13 +1,8 @@
-from dalex.explainer.explainer import Explainer
+from ._explainer.object import Explainer
 
-from dalex.instance_level.break_down import BreakDown
-
-from dalex.instance_level.shap import Shap
-
-from dalex.instance_level.ceteris_paribus import CeterisParibus
-
-from dalex.dataset_level.model_performance import ModelPerformance
-
-from dalex.dataset_level.variable_importance import VariableImportance
-
-from dalex.dataset_level.aggregated_profiles import AggregatedProfiles
+__all__ = [
+    "Explainer",
+    "dataset_level",
+    "instance_level",
+    "datasets"
+]
