+++ python/dalex/dalex/_explainer/__init__.py
@@ -1,0 +1,5 @@
+from .object import Explainer
+
+__all__ = [
+    "Explainer"
+]
