+++ python/dalex/dalex/_explainer/checks.py
