+++ python/dalex/dalex/dataset_level/__init__.py
@@ -1,3 +1,3 @@
-from .model_performance.model_performance import ModelPerformance
-from .variable_importance.variable_importance import VariableImportance
-from .aggregated_profiles.aggregated_profiles import AggregatedProfiles
+from ._model_performance.object import ModelPerformance
+from ._variable_importance.object import VariableImportance
+from ._aggregated_profiles.object import AggregatedProfiles
