+++ python/dalex/dalex/dataset_level/_aggregated_profiles/__init__.py
@@ -1,0 +1,5 @@
+from .object import AggregatedProfiles
+
+__all__ = [
+    "AggregatedProfiles"
+]
