+++ python/dalex/dalex/dataset_level/_model_performance/__init__.py
@@ -1,0 +1,5 @@
+from .object import ModelPerformance
+
+__all__ = [
+    "ModelPerformance"
+]
