+++ python/dalex/dalex/dataset_level/_variable_importance/__init__.py
@@ -1,0 +1,5 @@
+from .object import VariableImportance
+
+__all__ = [
+    "VariableImportance"
+]
