--- python/dalex/dalex/dataset_level/aggregated_profiles/__init__.py
@@ -1,1 +1,0 @@
-from .aggregated_profiles import AggregatedProfiles
