--- python/dalex/dalex/dataset_level/model_performance/__init__.py
@@ -1,1 +1,0 @@
-from .model_performance import ModelPerformance
