--- python/dalex/dalex/dataset_level/variable_importance/__init__.py
@@ -1,1 +1,0 @@
-from .variable_importance import VariableImportance
