--- python/dalex/dalex/explainer/__init__.py
@@ -1,1 +1,0 @@
-from .explainer import Explainer
