+++ python/dalex/dalex/instance_level/_ceteris_paribus/__init__.py
@@ -1,0 +1,5 @@
+from .object import CeterisParibus
+
+__all__ = [
+    'CeterisParibus'
+]
