+++ python/dalex/dalex/instance_level/_shap/__init__.py
@@ -1,0 +1,5 @@
+from .object import Shap
+
+__all__ = [
+    'Shap'
+]
