--- python/dalex/dalex/instance_level/break_down/__init__.py
@@ -1,1 +1,0 @@
-from .break_down import BreakDown
