--- python/dalex/dalex/instance_level/ceteris_paribus/__init__.py
@@ -1,1 +1,0 @@
-from .ceteris_paribus import CeterisParibus
