--- python/dalex/dalex/instance_level/shap/__init__.py
@@ -1,1 +1,0 @@
-from .shap import Shap
