+++ python/dalex/setup.py
@@ -1,13 +1,14 @@
 import setuptools
-
 from os import path
+
+
 this_directory = path.abspath(path.dirname(__file__))
 with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
 setuptools.setup(
     name="dalex",
-    version="0.1.0",
+    version="0.1.2",
     author="Wojciech Kretowicz, Hubert Baniecki, Przemyslaw Biecek",
     author_email="wojtekkretowicz@gmail.com, hbaniecki@gmail.com",
     description="DALEX in Python",
